% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgdm_parallelcrop.R
\name{bgdm_parallelcrop}
\alias{bgdm_parallelcrop}
\title{A function for cropping many large rasters in parallel to a specific area}
\usage{
bgdm_parallelcrop(
  raster_folder = "//esapov/esapov/ALL/Energy",
  shp_dsn,
  shp_layer,
  numCores = 15,
  out_folder
)
}
\arguments{
\item{raster_folder}{the filepath folder on local drive storing the set of rasters to be cropped}

\item{shp_dsn}{see dsn in sf::st_read()}

\item{shp_layer}{see layer in sf::st_read()}

\item{numCores}{the number of cores to extract each raster in parallel}

\item{out_folder}{the output folder for cropped rasters}
}
\description{
The function will crop large rasters in a specific location to the bounding box of a specific shapefile given a buffer area
}
