% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility_measurement_functions.R
\name{compute_networkaccess}
\alias{compute_networkaccess}
\title{A function to compute distance/cost to point of interest using cost-matrix system}
\usage{
compute_networkaccess(
  lines_obj,
  origins_dt,
  dest_dt,
  cost_matrix = FALSE,
  weight = "time",
  directed = FALSE,
  length_as_weight = TRUE
)
}
\arguments{
\item{lines_obj}{an \code{sf},\code{data.frame} or \code{sfnetwork} object}

\item{origins_dt}{an \code{sf} object showing the set of origin points}

\item{dest_dt}{on \code{sf} object showing the destimations or points of interest to optimize on}

\item{cost_matrix}{where or not the cost matrix should be returned}

\item{weight}{the variable within the \code{lines_obj} to be used as measure cost measure (e.g time)}

\item{directed}{FALSE if each edge is assumed to be two-way directed.}

\item{length_as_weight}{TRUE if \code{weight} is to be length of edges within the network}
}
\description{
A function to compute distance/cost to point of interest using cost-matrix system
}
