% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ntl.R
\name{get_daily_ntl}
\alias{get_daily_ntl}
\title{Download Daily Night Time Lights Data}
\usage{
get_daily_ntl(
  username,
  password,
  start_date,
  end_date,
  link_base = "https://eogdata.mines.edu/nighttime_light/nightly",
  satel_name,
  cloud_mask = T,
  cores = 1L
)
}
\arguments{
\item{username}{A character, username on NASA's Earth Observation Group database}

\item{password}{A character, password on NASA's Earth Observation Gruop database}

\item{start_date}{An object of class date, must be specified like as.Date("yyyy-mm-dd")}

\item{end_date}{An object of class date, must be specified like as.Date("yyyy-mm-dd")}

\item{link_base}{the link to the daily chirps data (no need to change this)}

\item{satel_name}{A character, for the satellite name, options are 'j01' and 'npp'}

\item{cloud_mask}{A logical, if TRUE the cloud masked data is pulled, otherwise rade9d data is downloaded}

\item{cores}{An integer, the number of cores to be used in parallel downloading}
}
\description{
Functions to download night time lights data from NASA's Earth Observation
Group's (EOG) database. The downloaded rasters are collected for different
time intervals, daily, monthly and annually.
}
