% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chirps.R
\name{get_sixhr_chirps}
\alias{get_sixhr_chirps}
\title{Download rainfall chirp data}
\usage{
get_sixhr_chirps(
  start_date,
  end_date,
  start_hr,
  end_hr,
  link_base = "https://data.chc.ucsb.edu/products/CHIRPS-2.0/",
  dsn = getwd(),
  cores = 1L
)
}
\arguments{
\item{start_date}{An object of class date, must be specified like as.Date("yyyy-mm-dd")}

\item{end_date}{An object of class date, must be specified like as.Date("yyyy-mm-dd")}

\item{start_hr}{An integer, the starting hour in the interval for pulling the data}

\item{end_hr}{An integer, the final hour in the interval for pulling the data}

\item{link_base}{the link to the daily chirps data (no need to change this)}

\item{dsn}{data source name (a local folder where the data is to be stored)}

\item{cores}{the number of PC cores to employ in pulling the data in parallel}

\item{resolution}{A character, gridded resolution of the data (options area "p05" i.e.
0.05 degree resolution and p25 for 0.25)}
}
\description{
Pull rainfall data from the CHIRPS data at different intervals (annual, daily, month, and six hourly)
The data is downloaded in raster format (.tif)
}
