% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goehelper_functions.R
\name{osm_processpoints}
\alias{osm_processpoints}
\title{This function cleans up OSM lines data in preparation for survey to survey imputation analysis}
\usage{
osm_processpoints(shapefile_path, osm_path, geoid_var, feature_var)
}
\arguments{
\item{shapefile_path}{filepath of country/region shapefile with polygons of interest}

\item{osm_path}{filepath to open street map lines data (road networks)}

\item{geoid_var}{the variable that points to the common identifier ID between shapefile_path object and osm_path object}

\item{feature_var}{specific feature of interest in the osm_path object}
}
\value{
A list containing two objects: a dataframe/datatable with original osm points data containing number of points
in polygon shapefile (from shapefile_path argument) and a lazy datatable version of the full osm points data
with point information
}
\description{
This function cleans up OSM lines data in preparation for survey to survey imputation analysis
}
