% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goehelper_functions.R
\name{saeplus_calibratepovrate}
\alias{saeplus_calibratepovrate}
\title{Benchmark Small Area Estimates}
\usage{
saeplus_calibratepovrate(
  hh_dt,
  pop_dt,
  ebp_obj,
  area_id,
  pop_var,
  harea_id,
  povline,
  welfare,
  weight,
  excl_outsample = TRUE
)
}
\arguments{
\item{hh_dt}{a household data.frame/data.table object containing area level IDs for which poverty rates were small area
estimated}

\item{pop_dt}{a population level data.frame/data.table object containing geospatial polygons and areas level IDs used in hh_dt
for which population level estimates can be computed}

\item{ebp_obj}{an object of class "emdi" "ebp" obtained from small area estimation containing area level poverty rates.
Area Level IDs must be matching across pop_dt, hh_dt and ebp_obj}

\item{area_id}{the level at which the poverty areas were estimated}

\item{harea_id}{a higher level at which intermediate SAE estimates will be computed}

\item{povline}{national poverty line to be used with the hh_dt object to determine survey level poverty rates}

\item{weight}{household level weights within the hh_dt object}

\item{excl_outsample}{whether or not outsample grids should be excluded in the benchmarking exercise}
}
\description{
This function benchmarks the SAE model estimates to match survey estimates at level of chosen level of
representativeness.
}
