% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility_measurement_functions.R
\name{clean_osmlines}
\alias{clean_osmlines}
\title{A function to clean and prepare geospatial lines (road network) data for measuring cost matrices}
\usage{
clean_osmlines(streets_obj)
}
\description{
A function to clean and prepare geospatial lines (road network) data for measuring cost matrices
}
